requirements.txt
google-genai
python-dotenv
requests
pydantic
beautifulsoup4
edge-tts
python-telegram-bot
praw
apscheduler
flask
gunicorn
